change setup file

+++ deploy/setup.sh
@@ -8,10 +8,11 @@
 GIT_URL=https://github.com/MyEtherWallet/simplex-api.git
 DOCKER_COMPOSE_VERSION=1.24.0
 ENV_FILE='.env'
+CURRENCY_FILE='currencyConfig.js'
 
 # GIT options
 # Use a branch other than master
-FROM_BRANCH=false
+FROM_BRANCH=true
 # The name of the branch to use
 BRANCH_NAME=add-cad-and-jpy;
 
@@ -21,7 +22,7 @@ STOP_DOCKER='false'
 FLAGGED='false'
 PURGE_DOCKER='false'
 REBUILD_RESTART='false'
-
+RUN_ALL='false'
 
 POSITIONAL=()
 while [[ $# -gt 0 ]]
@@ -54,7 +55,8 @@ case $key in
     shift # past argument
     ;;
     -a|--all)
-    DEFAULT=YES
+    RUN_ALL='true'
+    FLAGGED='true'
     shift # past argument
     ;;
     --default)
@@ -76,6 +78,7 @@ echo " -r | --restart : stop docker and run docker-compose "
 echo " -s | --stop-docker : stop all docker containers"
 echo " -p | --purge-docker : stop and remove all docker containers"
 echo " -b | --rebuild-restart-docker : remove docker containers, rebuild and run docker-compose"
+echo " -a | --all : run total setup or re-setup without asking for abort"
 echo "Running with no arguments initiates total setup or re-setup"
 echo "Note: total setup/re-setup does not replace an existing database data directory."
 
@@ -135,6 +138,11 @@ fi
 if [ "$PURGE_DOCKER" == 'true' ]; then
   echo "Stopping and removing all docker containers"
   purgeDocker
+fi
+
+if [ "$RUN_ALL" == 'true' ]; then
+  echo "Stopping and removing all docker containers"
+  doSetup
 fi
 
 if [ "$FLAGGED" == 'true' ]; then
@@ -228,6 +236,7 @@ purgeDocker(){
 
 buildDockerImages(){
     echo $PWD
+    cp ../${CURRENCY_FILE} ./
     cp ../${ENV_FILE} ./
     cd api;
     cp ../${ENV_FILE} ./
@@ -249,20 +258,24 @@ createDataDirectory(){
 }
 
 doSetup(){
-  echo "env file exists"
-  createDataDirectory
-  if [ -d "simplex-api" ]; then
-    purgeDocker
-    echo "prior simplex-api dir exists"
-    rm -rf ./simplex-api/
-    checkoutRepo
-    buildDockerImages
-    sudo docker-compose up -d --remove-orphans
-  else
-    echo "prior simplex-api dir does not exist"
-    checkoutRepo
-    buildDockerImages
-    sudo docker-compose up -d --remove-orphans
+  if [ -f ${ENV_FILE} ]; then
+    echo "env file exists"
+    createDataDirectory
+    if [ -d "simplex-api" ]; then
+      purgeDocker
+      echo "prior simplex-api dir exists"
+      rm -rf ./simplex-api/
+      checkoutRepo
+      buildDockerImages
+      sudo docker-compose up -d --remove-orphans
+    else
+      echo "prior simplex-api dir does not exist"
+      checkoutRepo
+      buildDockerImages
+      sudo docker-compose up -d --remove-orphans
+    fi
+    else
+      echo "ERROR: failed to begin setup. .env file does not exist"
   fi
 }
 
@@ -272,11 +285,9 @@ installDocker
 installDockerCompose
 
 runFromRepoDeploy
-if [ -f ${ENV_FILE} ]; then
+
 doSetup
-else
-    echo "ERROR: failed to begin setup. .env file does not exist"
-fi
+
